refactor: mover componentes grandes a archivos diferentes
